+++ app/src/app/[lng]/settings/account/index.tsx
