+++ app/src/app/[lng]/settings/project/InventoryView.tsx
