+++ app/src/app/[lng]/settings/project/deleteCityModal/index.tsx
